Large App Change
Added multiple files like Homescreen so im able to manage the code more effiicently as i added more functionality. 
The menu  items are interactive no so when a user clicks them it adds them to a basket 
Clicked menu item totals are added up and displayed. 
Added an Add Menu Item button 
Default menu items are also displayed so 2 for each category. 
And the colours for the app i made darker with black and grey